backend/pyproject.toml
[project]
name = "feedub-backend"
version = "1.0.0"
description = "Self-hosted Telegram message aggregation backend with AI insights"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "telethon>=1.30.3",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "pydantic>=2.5.0",
    "PyJWT>=2.8.0",
    "python-multipart>=0.0.6",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.1.1",
    "pydantic-settings>=2.11.0",
    "litellm>=1.40.0",
    "tenacity>=8.2.0",
    "cachetools>=5.3.0",
    "greenlet>=3.2.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.0,<1.0",
    "httpx>=0.25.2",
    "ruff>=0.1.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ty.environment]
# Python version and environment configuration
python-version = "3.12"
python = ".venv"
root = ["./src"]

[tool.ty.rules]
# Critical type safety errors (strict type checking)
possibly-unresolved-reference = "error"
invalid-argument-type = "error"
missing-argument = "error"
unsupported-operator = "error"
division-by-zero = "error"
possibly-unbound-attribute = "error"
unresolved-attribute = "error"
unknown-argument = "error"

# Code quality warnings
unused-ignore-comment = "warn"
redundant-cast = "warn"

[tool.ty.src]
# Files and directories to type check
include = ["src", "tests"]
respect-ignore-files = true

[tool.ty.terminal]
# Output configuration
output-format = "full"
error-on-warning = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--disable-warnings",
]
markers = [
    "unit: Unit tests",
    "contract: API contract tests",
    "integration: Integration tests requiring Docker",
    "asyncio: Async tests",
]

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
    "ty>=0.0.1a21",
    "types-PyJWT>=1.7.1",
    "ruff>=0.1.8",
    "pytest-asyncio>=0.23,<1.0",
]
